Refactor SupervisorJobs into a DataUpdateCoordinator
SupervisorJobs previously maintained its own job cache and lifecycle
using a manual dispatcher subscription and a refresh_data() method
called directly by HassioMainDataUpdateCoordinator.

Refactor it to be a proper DataUpdateCoordinator[dict[UUID, Job]]
alongside the other hassio coordinators:

- _async_update_data() fetches and flattens the job cache from Supervisor
- _async_refresh_finished() registers the dispatcher listener after
  the first successful refresh (replacing the first_update flag)
- _async_refresh() is overridden to compute deltas and notify
  job subscribers whenever the polled cache changes
- WS job events call async_set_updated_data() to update the cache
  in-place; supervisor restart events call async_request_refresh()
- async_shutdown() disconnects the dispatcher; the coordinator
  shutdown itself is handled via config_entry.async_on_unload

JobSubscription and SupervisorJobs are moved from jobs.py into
coordinator.py (per the home-assistant-enforce-class-module rule)
and jobs.py is deleted. All callers updated accordingly.
