+++ tests/components/conftest.py
@@ -900,7 +900,7 @@ def supervisor_client() -> Generator[AsyncMock]:
             return_value=supervisor_client,
         ),
         patch(
-            "homeassistant.components.hassio.jobs.get_supervisor_client",
+            "homeassistant.components.hassio.coordinator.get_supervisor_client",
             return_value=supervisor_client,
         ),
         patch(
